pyproject.toml
[project]
name = "ai-crop-disease-detector"
version = "0.1.0"
description = "AI-powered crop disease detection with CNN/ViT models"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "streamlit (>=1.48.1,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"


[tool.poetry.group.onnx.dependencies]
torch = "^2.8.0"
torchvision = "^0.23.0"
tensorflow = "^2.20.0"
onnx = "^1.18.0"
onnxruntime = "^1.22.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
